Fixing circular dep
